Creates MDX Stories - moves theme components to stories directory - creates Colors MDX story - creates Typography MDX story - adds @storybook-addon/docs - reorganizes colors by alphabet
