+++ .storybook/main.js
@@ -7,6 +7,7 @@ module.exports = {
     "@storybook/addon-links",
     "@storybook/addon-essentials",
     "@storybook/addon-interactions",
+    "@storybook/addon-docs",
     "@chakra-ui/storybook-addon",
     'storybook-addon-designs',
     'storybook-dark-mode'
