+++ design/atoms/theme/foundations/colors/index.ts
@@ -1,24 +1,4 @@
 const colors = {
-  neutral: {
-    'light-gray': "#F5F5F5",
-    'medium-gray': "#B6B3AC"
-  },
-  gold: {
-    '300': '#FFD26A',
-    '500': '#FABD2E',
-    '600': '#AF8420'
-  },
-  'chocolate': {
-    '400': '#272520',
-    '500': '#36342f',
-  },
-  'drab': {
-    '500': '#746338'
-  },
-  'sand': {
-    '500': '#FFDA85',
-    '600': '#E5BE63'
-  },
   'black': {
     '100': '#606060',
     '200': '#484848',
@@ -30,11 +10,30 @@ const colors = {
     '700': '#161616',
     '800': '#0D0C0C',
   },
-  // @todo gradient colors missing naming
+  'chocolate': {
+    '400': '#272520',
+    '500': '#36342f',
+  },
+  'drab': {
+    '500': '#746338'
+  },
+  gold: {
+    '300': '#FFD26A',
+    '500': '#FABD2E',
+    '600': '#AF8420'
+  },
   gradients: {
-    'background': 'linear(180deg, chocolate.400 0%, #1B1A18 100%)',
-    'cards': 'linear(180deg, #403824 0%, chocolate.400 100%)',
-    'glow': 'linear(180deg, gold.500 0%, #FFD36E 100%)'
+    'background': 'linear-gradient(180deg, #272520 0%, #1B1A18 100%)',
+    'cards': 'linear-gradient(180deg, #403824 0%, #272520 100%)',
+    'glow': 'linear-gradient(180deg, #FABD2E 0%, #FFD36E 100%)'
+  },
+  neutral: {
+    'light-gray': "#F5F5F5",
+    'medium-gray': "#B6B3AC"
+  },
+  'sand': {
+    '500': '#FFDA85',
+    '600': '#E5BE63'
   },
   'utility': {
     'error-red': '#FF7373'
