+++ design/atoms/theme/stories/theme.stories.mdx
@@ -1,0 +1,23 @@
+<!-- theme.stories.mdx -->
+import { Canvas, Meta, Story } from '@storybook/addon-docs';
+import { Box, Text } from '@chakra-ui/react';
+
+import { Page } from "../../../sbComponents/Page";
+import colors from '../foundations/colors';
+import { ColorsDisplay } from './Colors'
+import { Typography } from './Typography'
+
+<Meta title="atoms/Theme" />
+
+<Canvas>
+  <Story name="Color">
+    <ColorsDisplay />
+  </Story>
+</Canvas>
+
+<Canvas>
+  <Story name="Typography">
+    <Typography />
+  </Story>
+</Canvas>
+
