--- design/atoms/theme/theme.stories.tsx
@@ -1,11 +1,0 @@
-import { Meta, Story } from "@storybook/react";
-import { Typography as Text } from "./Typography";
-
-const storyConfig: Meta = {
-  title: "atoms/Theme",
-}
-export default storyConfig
-
-
-export const Typography: Story = Text
-
