+++ package.json
@@ -27,6 +27,7 @@
     "storybook-dark-mode": "^1.1.0"
   },
   "dependencies": {
+    "@storybook/addon-docs": "^6.5.9",
     "classnames": "^2.3.1"
   }
 }
